llm-analytics/pydantic-ai-python/requirements.txt
posthog>=3.0.0
pydantic-ai[openai]>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
